+++ R/gseAnalyzer.R
@@ -45,7 +45,8 @@ gseDisease <- function(geneList,
 ##'
 ##' perform gsea analysis
 ##' @param geneList order ranked geneList
-##' @param organism one of "hsa" and "mmu"
+##' @param ont one of "HDO", "HPO" or "MPO"
+##' @param organism one of "hsa" and "mm"
 ##' @param exponent weight of each step
 ##' @param minGSSize minimal size of each geneSet for analyzing
 ##' @param maxGSSize maximal size of each geneSet for analyzing
@@ -60,6 +61,7 @@ gseDisease <- function(geneList,
 ##' @author Yu Guangchuang
 ##' @keywords manip
 gseDO <- function(geneList,
+                  ont = "HDO",
                   organism = "hsa",
                   exponent=1,
                   minGSSize = 10,
@@ -81,7 +83,7 @@ gseDO <- function(geneList,
                verbose           = verbose,
                seed              = seed,
                by                = by,
-               ontology          = "DO", 
+               ontology          = ont, 
                ...)
 
 }
@@ -154,73 +156,5 @@ gseDGN <- function(geneList,
                seed              = seed,
                by                = by,
                ontology          = "DisGeNET",
-               ...)
-}
-
-##' MPO Gene Set Enrichment Analysis
-##'
-##'
-##' perform gsea analysis
-##' @inheritParams gseDO
-##' @return gseaResult object
-##' @export
-##' @author Erqiang Hu
-##' @keywords manip
-gseMPO <- function(geneList,
-                   exponent=1,
-                   minGSSize = 10,
-                   maxGSSize = 500,
-                   pvalueCutoff=0.05,
-                   pAdjustMethod="BH",
-                   verbose=TRUE,
-                   seed=FALSE,
-                   by = 'fgsea',
-                   ...) {
-                   
-
-    gseDisease(geneList          = geneList,
-               exponent          = exponent,
-               minGSSize         = minGSSize,
-               maxGSSize         = maxGSSize,
-               pvalueCutoff      = pvalueCutoff,
-               pAdjustMethod     = pAdjustMethod,
-               verbose           = verbose,
-               seed              = seed,
-               by                = by,
-               ontology          = "MPO",
-               ...)
-}
-
-##' MPO Gene Set Enrichment Analysis
-##'
-##'
-##' perform gsea analysis
-##' @inheritParams gseDO
-##' @return gseaResult object
-##' @export
-##' @author Erqiang Hu
-##' @keywords manip
-gseHPO <- function(geneList,
-                   exponent=1,
-                   minGSSize = 10,
-                   maxGSSize = 500,
-                   pvalueCutoff=0.05,
-                   pAdjustMethod="BH",
-                   verbose=TRUE,
-                   seed=FALSE,
-                   by = 'fgsea',
-                   ...) {
-                   
-
-    gseDisease(geneList          = geneList,
-               exponent          = exponent,
-               minGSSize         = minGSSize,
-               maxGSSize         = maxGSSize,
-               pvalueCutoff      = pvalueCutoff,
-               pAdjustMethod     = pAdjustMethod,
-               verbose           = verbose,
-               seed              = seed,
-               by                = by,
-               ontology          = "HPO",
                ...)
 }